yield

+++ IndigoELN/src/main/java/com/epam/indigoeln/web/rest/dto/calculation/BasicBatchModel.java
@@ -130,7 +130,7 @@ public class BasicBatchModel {
         this.eq = eq;
     }
 
-    public double getYield() {
+    public long getYield() {
         return yield;
     }
